use lightning css instead of parcel
