+++ package.json
@@ -88,6 +88,7 @@
     "istanbul-lib-report": "3.0.0",
     "istanbul-reports": "3.1.5",
     "launch-editor": "2.6.0",
+    "lightningcss": "1.18.0",
     "pidtree": "0.6.0",
     "string-width": "5.1.2",
     "strip-ansi": "7.0.1",
