+++ packages/jsenv-plugin-bundling/src/jsenv_plugin_bundling.js
@@ -1,4 +1,4 @@
-import { bundleCss } from "./css/bundle_css.js"
+import { bundleCss } from "./css/css_bundling_postcss.js"
 import { bundleJsClassicWorkers } from "./js_classic_workers/bundle_js_classic_workers.js"
 import { bundleJsModules } from "./js_module/bundle_js_modules.js"
