+++ packages/jsenv-plugin-minification/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@jsenv/plugin-minification",
-  "version": "1.0.2",
+  "version": "1.1.0",
   "license": "MIT",
   "repository": {
     "type": "git",
@@ -24,8 +24,8 @@
     "test": "node --conditions=development ./scripts/test.mjs"
   },
   "dependencies": {
-    "@jsenv/ast": "1.4.10",
     "html-minifier": "4.0.0",
+    "lightningcss": "1.18.0",
     "terser": "5.16.1"
   },
   "peerDependencies": {
