--- packages/jsenv-plugin-minification/src/css/minify_css.js
@@ -1,9 +1,0 @@
-import { minifyWithParcel } from "@jsenv/ast"
-
-export const minifyCss = ({ cssUrlInfo, context }) => {
-  const { code, map } = minifyWithParcel(cssUrlInfo, context)
-  return {
-    content: String(code),
-    sourcemap: map,
-  }
-}
