--- src/plugins/transpilation/css_parcel/jsenv_plugin_css_parcel.js
@@ -1,18 +1,0 @@
-import { transpileWithParcel } from "@jsenv/ast"
-
-// https://github.com/parcel-bundler/parcel-css
-export const jsenvPluginCssParcel = () => {
-  return {
-    name: "jsenv:css_parcel",
-    appliesDuring: "*",
-    transformUrlContent: {
-      css: (urlInfo, context) => {
-        const { code, map } = transpileWithParcel(urlInfo, context)
-        return {
-          content: String(code),
-          sourcemap: map,
-        }
-      },
-    },
-  }
-}
